requirements.txt
Django==2.2.1
pytz==2019.1
sqlparse==0.3.0
django-redis==4.10.0
djangorestframework==3.9.3
django-extensions==2.1.6
django-filter==2.1.0
redis==3.2.1
urllib3==1.24.3
gunicorn==19.9.0
gevent==1.3.7
qiniu==7.2.2
django-crispy-forms==1.7.2
psycopg2==2.8.2
django-mptt==0.10.0
Pillow==6.0.0
aliyun-python-sdk-core-v3==2.13.3
weixin-python==0.5.5
python-alipay-sdk==1.9.1
django-ckeditor==5.6.1